atomisp2: css2400 for irci_master_20150107_1500
Merge CSS2.1 for css2400  for
irci_master_20150107_1500 firmware

Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-16283
Reviewed-on: https://icggerrit.ir.intel.com/35410
Approver: Wei Tao <wei.tao@intel.com>

+++ drivers/external_drivers/camera/drivers/media/pci/atomisp2/css2400b0_v21/runtime/isp_param/interface/ia_css_isp_param_types.h
@@ -46,8 +46,8 @@ enum ia_css_param_class {
 
 /** ISP parameter descriptor */
 struct ia_css_isp_parameter {
-	uint16_t offset; /* Offset in isp_<mem>)parameters, etc. */
-	uint16_t size;   /* Disabled if 0 */
+	uint32_t offset; /* Offset in isp_<mem>)parameters, etc. */
+	uint32_t size;   /* Disabled if 0 */
 };
